pyproject.toml
[project]
name = "rag-pgvector-tuning"
version = "0.1.0"
description = "Learn RAG retrieval with PostgreSQL pgvector: ingest, ANN search, and YAML-driven HNSW/IVFFlat runtime tuning (FastAPI demo, no API keys)."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.109.0",
  "uvicorn[standard]>=0.27.0",
  "asyncpg>=0.29.0",
  "pydantic>=2.5.0",
  "pydantic-settings>=2.1.0",
  "pyyaml>=6.0",
  "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.26.0",
  "pytest>=8.0.0",
  "pytest-asyncio>=1.3.0",
  "pytest-cov>=7.1.0",
]

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "httpx>=0.26.0",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "pytest>=8.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["src", "."]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "integration: requires PostgreSQL with pgvector (see tests/test_integration_api.py)",
]

[tool.coverage.run]
branch = true
source = ["rag"]

[tool.coverage.report]
fail_under = 80
show_missing = true
